Move GameStateChanger from lib to core
