+++ roguelike-game/src/main/java/models/ConfusedMob.java
@@ -112,7 +112,6 @@ public class ConfusedMob extends Mob {
     private boolean tryMakeMove(Mob player, Room room, int dx, int dy) {
         if (player.posX() == mob.posX() + dx && player.posY() == mob.posY() + dy) {
             player.stepOn(mob);
-            System.out.println("aboba");
             return true;
         }
