+++ roguelike-game/src/main/java/models/PassiveMobBehavior.java
@@ -2,6 +2,5 @@ package models;
 
 public class PassiveMobBehavior implements MobBehavior {
     @Override
-    public void makeNextMove(Mob player, Mob mob, Room room) {
-    }
+    public void makeNextMove(Mob player, Mob mob, Room room) {}
 }
